fix: fix spaced-repetitions logic
