+++ src/components/main/components/mainStorage/mainStorage.js
@@ -1,10 +1,11 @@
 import mainController from '../controller/main.controller';
-// import spacedRepetitions from '../spacedRepetitions/spacedRepetitions';
+import spacedRepetitions from '../spacedRepetitions/spacedRepetitions';
 
 import { checkUserInfo } from '../../../../services/common/services.common.api_service.helper';
 
 import { EMPTY } from '../../../../common/common.constants';
 import { MINI_GAMES_NAMES } from '../../common/main.constants';
+import { WORD_CATEGORY_TO_INDEX } from '../spacedRepetitions/common/constants';
 
 class MainStorage {
   constructor() {
@@ -76,7 +77,6 @@ class MainStorage {
   loadMiniGamesResults() {
     try {
       const { userId } = checkUserInfo();
-      // const loadedMiniGamesResults = JSON.parse(localStorage.getItem(`miniGamesResults-${userId}`));
       this.miniGamesResults = JSON.parse(localStorage.getItem(`miniGamesResults-${userId}`)) || this.miniGamesResults;
     } catch (error) {
       console.info(error.message);
@@ -88,17 +88,32 @@ class MainStorage {
     window.removeEventListener('beforeunload', this.beforeUnloadHandler);
   }
 
-  async getWordsToLearn() {
-    mainController.spinner.show();
-    this.wordsToLearn = await mainController.getAllUserAggregatedWords({ group: '3', wordsPerPage: '27' });
-    mainController.spinner.hide();
-    return this.wordsToLearn;
+  getUserEnglishLevel() {
+    return mainController.englishLevel;
   }
 
-  init() { // async?
+  getWordsToLearn() {
+    spacedRepetitions.userWordsCollection.sort(
+      (wordDataA, wordDataB) => WORD_CATEGORY_TO_INDEX[wordDataA.userWord.difficulty]
+        - WORD_CATEGORY_TO_INDEX[wordDataB.userWord.difficulty],
+    );
+
+    spacedRepetitions.userWordsCollection.sort(
+      (wordDataA, wordDataB) => {
+        if (wordDataA.userWord.difficulty === wordDataB.userWord.difficulty) {
+          return wordDataA.userWord.optional.repeatDate
+            - wordDataB.userWord.optional.repeatDate;
+        }
+        return false;
+      },
+    );
+
+    return spacedRepetitions.userWordsCollection;
+  }
+
+  init() {
     this.loadMiniGamesResults();
     window.addEventListener('beforeunload', this.beforeUnloadHandler);
-    // return this.wordsToLearn;
   }
 }
