+++ src/components/main/components/spacedRepetitions/common/constants.js
@@ -1,0 +1,18 @@
+export const WORD_CATEGORY_TO_INDEX = {
+  fetched: 0,
+  new: 1,
+  hard: 2,
+  normal: 3,
+  good: 4,
+  excellent: 5,
+  learned: 6,
+};
+
+export const INDEX_TO_CATEGORY = {
+  0: 'new',
+  1: 'hard',
+  2: 'normal',
+  3: 'good',
+  4: 'excellent',
+  5: 'learned',
+};
