+++ src/components/main/main.app.js
@@ -1,6 +1,4 @@
 import { mainStorage } from './components/mainStorage/mainStorage';
-// import mainController from './components/controller/main.controller';
-// import spacedRepetitions from './components/spacedRepetitions/spacedRepetitions';
 import router from './components/router/router';
 import './scss/main.styles.scss';
