+++ src/components/speakit/components/model/model.js
@@ -123,6 +123,9 @@ class Model {
   }
 
   init() {
+    console.log(mainStorage.getWordsToLearn());
+    console.log(mainStorage.getUserEnglishLevel());
+    debugger;
     this.loadCurrentResults();
     this.loadLongResults();
     this.wordsAPI = new WordsApi();
